Add commands to allow user to specify target room
This commit introduces a new variant on the 'ignore me' and 'listen to
me' commands by adding support for specifying the room in the command
(e.g., 'ignore me in #some-room'). In adapters where the user can speak
with the robot in a "room" directly - like the Slack adapter - this will
allow the user to privately set ignore and listen rules.

Fixes #2